refactor(orm): make the code more logical
